test: event flow based test refactorized
